requirements.txt
absl-py==0.14.0
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==1.1.9
APScheduler==3.6.3
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cached-property==1.5.2
cachetools==4.2.2
Cap1xxx==0.1.3
cbor2==5.4.1
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
Deprecated==1.2.13
devicehive==2.1.6
devicehive-webconfig==1.0.2
docutils==0.14
dominate==2.6.0
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-WTF==0.15.1
fourletterphat==0.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
gpiozero==1.6.2
greenlet==1.1.2
grpcio==1.40.0
h5py==3.4.0
html5lib==1.0.1
idna==2.6
importlib-metadata==4.8.1
imutils==0.5.4
isort==4.3.4
itsdangerous==2.0.1
jedi==0.13.2
Jinja2==3.0.2
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
luma.core==2.3.1
luma.oled==3.8.1
lxml==4.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mpu6050-raspberrypi==1.1
mypy==0.670
mypy-extensions==0.4.1
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
opencv-contrib-python==3.4.3.18
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
protobuf==3.18.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybase64==1.2.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.53.3
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
pyxdg==0.25
pyzmq==22.2.1
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.26.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
rsa==4.7.2
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
smbus2==0.4.1
sn3218==1.2.7
soupsieve==1.8
SpeechRecognition==3.8.1
spidev==3.5
SQLAlchemy==1.4.25
ssh-import-id==5.7
thonny==3.3.10
tornado==6.1
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
unicornhathd==0.0.4
urllib3==1.24.1
visitor==0.1.3
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.2
wrapt==1.10.11
WTForms==2.3.3
zipp==3.5.0
zmq==0.0.0
